Remove announcement sub type for government_document_supertype
We've done the work to remove the Unknown NewsArticleType from Whitehall. This type was used for uncategorised NewsArticles, which have now been turned into NewsStories or PressReleases.
The Unknown Type would default to a document_type of 'announcement' which was quite confusing. Since we've removed all data for the Unknown NewsArticleTypes, this is not used any more and can be removed from this repo.