Add email_document_supertype > announcements > world_news_story
The subscriber lists in email-alert-api have an
email_document_supertype of announcements, but
we're sending this through as 'other' for these
document types. This means we don't match those
subscriber lists and some people wouldn't get
notified if we switched to sending through
email-alert-api instead of govuk-delivery. This
fixes that.