Add new content purpose sub and supergroups
As part of building the navigation topic pages we did work using the
JTBD framework to identify new content purpose sub and supergroups.
In order to help users find the content they require, we can group
document types by the needs they're designed to meet and the purpose of
the content.
See the wiki for more info:
https://gov-uk.atlassian.net/wiki/spaces/GFED/pages/305201156/Document+type+groupings

Due to the fact that the content purpose supergroups group subgroups not
document types, we have added a new yaml file for the supergroups.

The data lint specs have been change to test both data files.