Change supergroup documentation in yaml file.
This documentation will be imported into and used in the developer docs
so I needed to change some of the links and the markdown syntax for the
HTML to display correctly.

This documentation will appear on the Content purpose supergroup page in
the documentation. https://docs.publishing.service.gov.uk/document-types/content_purpose_supergroup.html