Find document_types in a supertype group
We want to decouple content_purpose supertypes from having to be indexed prior
to being useful, as we want to be able to iterate them easily.

They are mostly used in applications that use filter_content_purpose_supergroup
type queries to rummager to return the required content.  To remove the
requirement for rummager to be up to date, we'll allow applications to resolve
a list of document types that belong to a supertype.  They can then generate
queries using `filter_content_store_document_type` and the list of document
types.