Update readme
Content pages and nav team need to iterate supertypes.  Republishing content
after changes takes too long and may have contributed to recent incidents so
should be minimised. In addition to this, changing the id of a supertype group
is really tricky.

By allowing this gem to be used in other applications, we can circument the
above issues.