Add translations to content_purpose supertypes
We want to be able to trivially change the title of a supertype and keep it
consistent between applications.  The initial purpose is to change "Policy and
engagement" to "Policy papers and consultations".

We've also added translations for some subgroups that don't exist yet - they'll
be added next.

The Railtie loads translation files from the gem in apps that consume it.  This
won't happen in apps without Rails.

[Supertypes v2.2](https://docs.google.com/spreadsheets/d/1n9OJ4s7AOd5-JDOz_gXgWLdLGxzzzJUxnc-bJXOyQQ0/edit#gid=1065921575)

https://trello.com/c/Q88UJzp5/177-update-collections-to-use-the-govukdocumenttypes-gem