Release version 1.0.0
Includes:

Remove deprecated supertypes
Add Flood & Coastal Erosion Risk Management Report (FCERM) Research Report to research supergroup

The removal of the old types is a breaking change (if they're used), though
we're pretty confident that they aren't used anywhere.

Releasing v1 as we've used this gem for ages, and it's been stable for a long
time.