Use `main` branch for Content Schemas
The CI pipeline was checking out the `deployed-to-production` branch of the Publishing API repository to lint the document types against defined content schemas.

However, the `deployed-to-production` branch no longer exists in GOV.UK repositories. This is an implementation detail from the old EC2-based hosting platform, but the new EKS hosting environment simply deploys the `main` branch without tagging releases.

We therefore need to switch this repository's CI pipeline over to use the `main` branch of Publishing API, so it picks up the current version of the GOV.UK content schemas.